# ait-core-soriano/docs/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AIT-CORE Soriano Insurance Platform API
  version: 1.0.0
  description: |
    # AIT-CORE Soriano Insurance Platform API

    Comprehensive RESTful API for managing insurance operations, policies, claims, and customer relationships.

    ## Features

    - **Authentication & Authorization**: JWT-based authentication with role-based access control
    - **User Management**: Complete user lifecycle management
    - **Insurance Operations**: Policies, quotes, claims, and underwriting
    - **Real-time Communication**: WebSocket support for live updates
    - **Analytics**: Advanced reporting and business intelligence
    - **Notifications**: Multi-channel notification system (Email, SMS, Push)
    - **Document Management**: Secure document storage and retrieval
    - **Audit Trail**: Comprehensive activity logging

    ## Versioning

    This API uses URI versioning (e.g., /api/v1/users)

    ## Authentication

    Use the `/auth/login` endpoint to obtain a JWT token. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```

    ## Rate Limiting

    - Anonymous: 100 requests per hour
    - Authenticated: 1000 requests per hour
    - Premium: 10000 requests per hour

    ## Support

    - **Email**: support@soriano.com
    - **Documentation**: https://docs.soriano.com
    - **Status Page**: https://status.soriano.com

  contact:
    name: AIT-CORE Support
    url: https://ait-core.soriano.com
...

  license:
...

...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication

    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Refresh token stored in HTTP-only cookie
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2024-01-28T12:00:00Z'
        requestId:
          type: string
          example: req_abc123xyz
...
    User:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - role
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          example: user@example.com
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        phone:
          type: string
          example: '+34612345678'
        role:
          type: string
          enum: [USER, MANAGER, ADMIN, SUPER_ADMIN]
          example: USER
        isActive:
          type: boolean
          example: true
        emailVerified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Obtain a new access token using refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: Get the profile of the currently authenticated user
      operationId: getCurrentUser
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
...